fix(goamiddleware): refactor goa middleware code and add licenses
