+++ goamiddleware/jwt_token_context.go
@@ -1,7 +1,28 @@
+// The MIT License (MIT)
+
+// Copyright (c) 2015 Raphael Simon and goa Contributors
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
 package goamiddleware
 
 import (
-	"crypto/ecdsa"
 	"crypto/rsa"
 	"fmt"
 	"net/http"
@@ -45,21 +66,21 @@ func TokenContext(validationKeys interface{}, validationFunc goa.Middleware, sch
 				)
 
 				if len(rsaKeys) > 0 {
-					token, err = parseRSAKeys(rsaKeys, "RS", incomingToken)
+					token, err = validateRSAKeys(rsaKeys, "RS", incomingToken)
 					if err == nil {
 						parsed = true
 					}
 				}
 
 				if !parsed && len(ecdsaKeys) > 0 {
-					token, err = parseECDSAKeys(ecdsaKeys, "ES", incomingToken)
+					token, err = validateECDSAKeys(ecdsaKeys, "ES", incomingToken)
 					if err == nil {
 						parsed = true
 					}
 				}
 
 				if !parsed && len(hmacKeys) > 0 {
-					token, err = parseHMACKeys(hmacKeys, "HS", incomingToken)
+					token, err = validateHMACKeys(hmacKeys, "HS", incomingToken)
 					if err == nil {
 						parsed = true
 					}
@@ -75,49 +96,4 @@ func TokenContext(validationKeys interface{}, validationFunc goa.Middleware, sch
 			return nextHandler(ctx, rw, req)
 		}
 	}
-}
-
-func parseRSAKeys(rsaKeys []*rsa.PublicKey, algo, incomingToken string) (token *jwt.Token, err error) {
-	for _, pubkey := range rsaKeys {
-		token, err = jwt.Parse(incomingToken, func(token *jwt.Token) (interface{}, error) {
-			if !strings.HasPrefix(token.Method.Alg(), algo) {
-				return nil, goajwt.ErrJWTError(fmt.Sprintf("unexpected signing method: %v", token.Header["alg"]))
-			}
-			return pubkey, nil
-		})
-		if err == nil {
-			return
-		}
-	}
-	return
-}
-
-func parseECDSAKeys(ecdsaKeys []*ecdsa.PublicKey, algo, incomingToken string) (token *jwt.Token, err error) {
-	for _, pubkey := range ecdsaKeys {
-		token, err = jwt.Parse(incomingToken, func(token *jwt.Token) (interface{}, error) {
-			if !strings.HasPrefix(token.Method.Alg(), algo) {
-				return nil, goajwt.ErrJWTError(fmt.Sprintf("unexpected signing method: %v", token.Header["alg"]))
-			}
-			return pubkey, nil
-		})
-		if err == nil {
-			return
-		}
-	}
-	return
-}
-
-func parseHMACKeys(hmacKeys [][]byte, algo, incomingToken string) (token *jwt.Token, err error) {
-	for _, key := range hmacKeys {
-		token, err = jwt.Parse(incomingToken, func(token *jwt.Token) (interface{}, error) {
-			if !strings.HasPrefix(token.Method.Alg(), algo) {
-				return nil, goajwt.ErrJWTError(fmt.Sprintf("unexpected signing method: %v", token.Header["alg"]))
-			}
-			return key, nil
-		})
-		if err == nil {
-			return
-		}
-	}
-	return
 }
